use updated code for ios and android
